Bugfix. Files did not get deleted during import.
Fixed a nasty bug which prevented deleted files actually getting
deleted during the import process. So, if a file foo.cs would have
been deleted in TFS, it still would be present in the converted Git
repository.

Usually this would not be a problem because the corresponding .csproj
should be updated to *not* build that file.
That's why I didn't detect this bug for so long ;-(